+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file.
  
 ## [Unreleased]
+#### Added
+- Added some methods for getting values to Adapter.php
+- Added fallback to RPC for methods we're not able to run in LDAP
+
 #### Changed
+- Changed the way of getting attribute names for interfaces: through internal attribute names in perun_attributes.php config
 - Return sorted eduPersonEntitlement
 
 ## [v3.9.0]